The progress bar of next song is not properly started in offload use case
The gapless data was not passed to HAL, which in turn to DRV/FW.
During start of every song gapless information is passed to
DSP to indicate the reset of hardware pointer. This pointer
is used in getting the render position. If some files does
not have gapless header, this value is not passed to DSP.
This causes wrong updation of render position for the next song
in the loop mode. This change will update the necessary information
to DSP appropriately.

Change-Id: I3dc71f6fb38e7067610813503ff60340e7dcb815
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>